build rpc test container image

+++ scripts/deploy-docker.sh
@@ -3,9 +3,18 @@
 dockerComposeFilePath="deployments/docker-compose"
 
 mkdir -p ${dockerComposeFilePath}/configs
-cp configs/serverNameExample.yml ${dockerComposeFilePath}/configs
+if [ ! -f "${dockerComposeFilePath}/configs/serverNameExample.yml" ];then
+  cp configs/serverNameExample.yml ${dockerComposeFilePath}/configs
+fi
 
 # shellcheck disable=SC2164
 cd ${dockerComposeFilePath}
 
+if [ "$1"x = "stop"x ] ;then
+  docker-compose down
+  exit 0
+fi
+
 docker-compose up -d
+
+echo "path is 'deployments/docker-compose'"
