+++ scripts/image-build.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
-# Image built by directly copying the compiled binaries. Advantage: fast build, disadvantage: the image is twice as large as a two-stage build.
+# build the docker image using the binaries, if you want to reduce the size of the image,
+# use upx to compress the binaries before building the image.
 
 serverName="serverNameExample_mixExample"
 # image name of the service, no capital letters
@@ -31,9 +32,20 @@ CONFIG_PATH="configs"
 # only grpc use start
 bash scripts/grpc_health_probe.sh
 # only grpc use end
-CGO_ENABLED=0 GOOS=linux GOARCH=amd64 GOPROXY=https://goproxy.cn,direct go build -gcflags "all=-N -l" -o ${BIN_FILE} cmd/${serverName}/*.go
+CGO_ENABLED=0 GOOS=linux GOARCH=amd64 GOPROXY=https://goproxy.cn,direct go build -o ${BIN_FILE} cmd/${serverName}/*.go
 mv -f ${BIN_FILE} ${DOCKERFILE_PATH}
 mkdir -p ${DOCKERFILE_PATH}/${CONFIG_PATH} && cp -f ${CONFIG_PATH}/${serverName}.yml ${DOCKERFILE_PATH}/${CONFIG_PATH}
+
+# compressing binary file
+#cd ${DOCKERFILE_PATH}
+#upx -9 ${serverName}
+#upx -9 grpc_health_probe
+#cd -
+
 echo "docker build -f ${DOCKERFILE} -t ${IMAGE_NAME_TAG} ${DOCKERFILE_PATH}"
 docker build -f ${DOCKERFILE} -t ${IMAGE_NAME_TAG} ${DOCKERFILE_PATH}
+
+if [ X"${serverName}" = X ];then
+        exit 0
+fi
 rm -rf ./${DOCKERFILE_PATH}/${serverName} ${DOCKERFILE_PATH}/configs ${DOCKERFILE_PATH}/grpc_health_probe
